requirement.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.13
django-cors-headers==3.1.1
django-rest-registration==0.5.1
django-rest-swagger==2.2.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.5.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
PyJWT==1.7.1
pytz==2020.5
requests==2.25.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
